feat: adds threading and synchronization fixes for savingm, receiving and iterating messages

+++ client/patient_walkietalkie.py
@@ -9,7 +9,7 @@ from appJar import gui
 from recognizer import get_state_machine
 from recorder import Recorder
 from stmpy import Driver, Machine
-from threading import Thread
+from threading import Thread, Lock
 from tts import Speaker
 from uuid import uuid4
 import stmpy
@@ -61,6 +61,7 @@ class WalkieTalkie:
         self.debug = debug
         self.app = None
         self.message_in_queue = False
+        self.lock = Lock()
 
         self.recorder = Recorder(self)
         self.text_to_speech = Speaker()
@@ -181,7 +182,8 @@ class WalkieTalkie:
         elif payload.get('data'):
             self.stm.send('replay_save_message', args=[payload])
 
-    def save_message(self, payload):
+    def threaded_save(self, lock, payload):
+        lock.acquire()
         try:
             sender_name = payload.get('device_owner_name_from')
             self.tts(f"Received message from {sender_name}")
@@ -190,13 +192,19 @@ class WalkieTalkie:
             data = base64.b64decode(wf)
             # Get queue length and saves message in the FIFO order
             queue_number = len(os.listdir("message_queue"))
+            print("QUEUE", queue_number)
             with open(f'message_queue/{queue_number}.wav', 'wb') as fil:
                 fil.write(data)
                 self._logger.debug(f'Message saved to /message_queue/{queue_number}.wav')
-            self.update_led(False)
-            self.iterate_queue(False)
         except:
             self._logger.error(f'Payload could not be read!')
+        lock.release()
+        self.iterate_queue(False)
+
+    def save_message(self, payload):
+        th = Thread(target=self.threaded_save, args=[self.lock,payload])
+        th.start()
+        th.join()
 
     def play_replay_message(self, payload):
         try:
@@ -237,17 +245,27 @@ class WalkieTalkie:
             return True
         return False
 
-    def iterate_queue(self, remove = True):
+    def threaded_iterate(self, lock, remove):
+        lock.acquire()
         queue_folder = "message_queue"
         num = 1
-        for filename in os.listdir(queue_folder):
-            if filename.split(".")[0] == "1" and remove:
+        listdir = os.listdir(queue_folder)
+        listdir.sort()
+        for filename in listdir:
+            if filename.split(".")[0] == "1" and num == 1 and remove:
                 os.remove(f"{queue_folder}/{filename}")
             else:
-                if filename != ".gitkeep":                
+                if filename != ".gitkeep":
+                    print(filename, num)
                     os.rename(f"{queue_folder}/{filename}", f"{queue_folder}/{num}.wav")
                     num += 1
         self.update_led(False)
+        lock.release()
+
+    def iterate_queue(self, remove = True):
+        th = Thread(target=self.threaded_iterate, args=[self.lock, remove]);
+        th.start()
+        th.join()
 
     # Request replay message from the server
     def get_latest_user_message(self):
